# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Affinity Risk Solutions for AAA Clubs - Telematics Insurance Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clubs/{clubId}:
    get:
      operationId: getClub
      tags: [clubs]
      summary: Get club by ID
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
  /clubs/{clubId}/summary:
    get:
      operationId: getClubSummary
      tags: [clubs]
      summary: Get risk summary for a club across all facilities
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubSummary"

  # ─── Facilities ───────────────────────────────────────────────────────────────
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List drivers, optionally filtered by facility
      parameters:
        - name: facilityId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
  /drivers/{driverId}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get driver by ID
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{driverId}/behavior:
    get:
      operationId: getDriverBehavior
      tags: [drivers]
      summary: Get detailed behavior analysis for a driver
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver behavior details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverBehavior"

  # ─── Vehicles ─────────────────────────────────────────────────────────────────
...
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List certificates of insurance, optionally filtered by facility
      parameters:
        - name: facilityId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certificate"

    post:
      operationId: uploadCertificate
      tags: [certificates]
      summary: Upload a new certificate of insurance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Certificate uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
  /certificates/{certificateId}:
    get:
      operationId: getCertificate
      tags: [certificates]
      summary: Get certificate by ID
      parameters:
        - name: certificateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"

  # ─── Accidents & Claims ───────────────────────────────────────────────────────
  /accidents:
    get:
      operationId: listAccidents
      tags: [accidents]
      summary: List accident alerts, optionally filtered by facility
      parameters:
        - name: facilityId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of accidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Accident"

    post:
      operationId: reportAccident
      tags: [accidents]
      summary: Report a new accident alert from telematics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccidentInput"
      responses:
        "201":
          description: Accident reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accident"
  /accidents/{accidentId}:
    get:
      operationId: getAccident
      tags: [accidents]
      summary: Get accident by ID
      parameters:
        - name: accidentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accident"
  /accidents/{accidentId}/initiate-claim:
    post:
      operationId: initiateClaim
      tags: [accidents]
      summary: Initiate a claim for an accident
      parameters:
        - name: accidentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimInput"
      responses:
        "200":
          description: Claim initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accident"

  # ─── Dashboard ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Club:
      type: object
      required: [id, name, state, region, totalFacilities, avgRiskScore, status]
      properties:
        id:
          type: integer
        name:
          type: string
        state:
          type: string
        region:
          type: string
        totalFacilities:
          type: integer
        avgRiskScore:
          type: number
        status:
          type: string
          enum: [active, review, suspended]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        createdAt:
          type: string
    ClubSummary:
      type: object
      required: [clubId, totalFacilities, totalDrivers, totalVehicles, avgRiskScore, highRiskCount, openAccidents, certificatesExpiringSoon]
      properties:
        clubId:
          type: integer
        totalFacilities:
          type: integer
        totalDrivers:
          type: integer
        totalVehicles:
          type: integer
        avgRiskScore:
          type: number
        highRiskCount:
          type: integer
        openAccidents:
          type: integer
        certificatesExpiringSoon:
          type: integer
...
    Driver:
      type: object
      required: [id, facilityId, firstName, lastName, licenseNumber, status, riskScore, riskTier, totalMiles, totalTrips]
      properties:
        id:
          type: integer
        facilityId:
          type: integer
        facilityName:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        licenseNumber:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, suspended]
        riskScore:
          type: number
        riskTier:
          type: string
          enum: [low, moderate, high, critical]
        totalMiles:
          type: number
        totalTrips:
          type: integer
        hireDate:
          type: ["string", "null"]
        createdAt:
          type: string
    DriverBehavior:
      type: object
      required: [driverId, riskScore, hardBrakingCount, harshAccelerationCount, speedingCount, phoneUsageCount, nightDrivingHours, avgSpeed, maxSpeed, totalMiles]
      properties:
        driverId:
          type: integer
        riskScore:
          type: number
        hardBrakingCount:
          type: integer
        harshAccelerationCount:
          type: integer
        speedingCount:
          type: integer
        phoneUsageCount:
          type: integer
        nightDrivingHours:
          type: number
        avgSpeed:
          type: number
        maxSpeed:
          type: number
        totalMiles:
          type: number
        trend:
          type: string
          enum: [improving, stable, worsening]
...
    Certificate:
      type: object
      required: [id, facilityId, policyNumber, insurer, coverageType, coverageAmount, effectiveDate, expirationDate, status]
      properties:
        id:
          type: integer
        facilityId:
          type: integer
        facilityName:
          type: ["string", "null"]
        policyNumber:
          type: string
        insurer:
          type: string
        coverageType:
          type: string
        coverageAmount:
          type: number
        effectiveDate:
          type: string
        expirationDate:
          type: string
        status:
          type: string
          enum: [current, expiring_soon, expired]
        fileUrl:
          type: ["string", "null"]
        uploadedAt:
          type: string
    CertificateInput:
      type: object
      required: [facilityId, policyNumber, insurer, coverageType, coverageAmount, effectiveDate, expirationDate]
      properties:
        facilityId:
          type: integer
        policyNumber:
          type: string
        insurer:
          type: string
        coverageType:
          type: string
        coverageAmount:
          type: number
        effectiveDate:
          type: string
        expirationDate:
          type: string
        fileUrl:
          type: ["string", "null"]
    Accident:
      type: object
      required: [id, vehicleId, driverId, facilityId, latitude, longitude, severity, status, alertedAt]
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehiclePlate:
          type: ["string", "null"]
        driverId:
          type: integer
        driverName:
          type: ["string", "null"]
        facilityId:
          type: integer
        facilityName:
          type: ["string", "null"]
        latitude:
          type: number
        longitude:
          type: number
        severity:
          type: string
          enum: [minor, moderate, major, total_loss]
        status:
          type: string
          enum: [alerted, claim_initiated, under_review, resolved]
        description:
          type: ["string", "null"]
        claimNumber:
          type: ["string", "null"]
        alertedAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
    AccidentInput:
      type: object
      required: [vehicleId, driverId, facilityId, latitude, longitude, severity]
      properties:
        vehicleId:
          type: integer
        driverId:
          type: integer
        facilityId:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        severity:
          type: string
          enum: [minor, moderate, major, total_loss]
        description:
          type: ["string", "null"]
    ClaimInput:
      type: object
      required: [claimNumber]
      properties:
        claimNumber:
          type: string
        notes:
          type: ["string", "null"]
...