# lib/api-zod/src/generated/types/club.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Affinity Risk Solutions for AAA Clubs - Telematics Insurance Platform
 * OpenAPI spec version: 0.1.0
 */
import type { ClubStatus } from './clubStatus';

export interface Club {
  id: number;
  name: string;
  state: string;
  region: string;
  totalFacilities: number;
  avgRiskScore: number;
  status: ClubStatus;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  createdAt?: string;
}